Implement a real user style sheet
With passing the `cssOrigin` property to `chrome.tabs.insertCSS`
method's second argument. However, the problem is that it is taking a
while before our style sheet loads. That is, there is a flash of the old
style before our style sheet loads. This did not happen when I was using
the "content_scripts" property in "manifest.json". However, that did not
allow me to inject the style sheet with "user origin". It let me inject
only as "author origin".